requirements.txt
oauth2client
gspread
attrs

pytest

ipython
pdbpp
